Ajuste en Arbol de archivos v2
