+++ src/components/Cards/AppCardProductDetail.jsx
@@ -1,21 +1,22 @@
 import React from "react";
 
 // import AppCheckboxFilter from "../components/commons/AppCheckboxFilter";
-import AppImage from "./AppImage";
-import AppTextarea from "./AppTextarea";
+import AppImage from "../commons/AppImage";
+import AppTextarea from "../commons/AppTextarea";
 import AppPostAuthorInfo from "./AppPostAuthorInfo";
+import AppPostDateCreation from "./AppPostDateCreation";
 import AppPostReactions from "./AppPostReactions";
 
-import Styles from "./AppCardArticleDetail.module.css"
+import Styles from "./AppCardProductDetail.module.css"
 
-import banner from "../../img/card/banner-article.jpg"
+import banner from "../../img/card/banner-product.jpg"
 
-export default function AppCardArticleDetail(props) {
+export default function AppCardProductDetail() {
 
   return (
     <React.Fragment>
 
-    <article className="card mb-3 nav-view-post">
+    <article className="card mb-3">
 
         <AppImage classImage="card-img-top w-100" pathImage={banner} altImage="banner-img"></AppImage>
 
@@ -29,16 +30,27 @@ export default function AppCardArticleDetail(props) {
             {/* Tags */}
             <div className={`${Styles.PublishedTags} tags-color d-flex flex-wrap`}> #webdev #wordpress #frontend #tutorial </div>
 
-            {/* Contenedor del avatar y el name */}
-            <AppPostAuthorInfo estado="full-info"/>
 
             {/* Content */}
             <div className={`${Styles.PublishedContent}`}>
                 <p>Lorem ipsum dolor, sit amet consectetur adipisicing elit. Molestiae asperiores perferendis beatae perspiciatis facilis! Rem, mollitia nesciunt assumenda vel magni doloremque fuga iste quis sunt sequi expedita ratione quos laborum.</p>
             </div>
 
-            {/* Reactions */}
-            <AppPostReactions/>
+            {/* Reactions and Creation Date */}
+            <div className={`py-2`}>
+
+                <div className="row d-flex align-items-center">
+                    <div className="col-12 col-lg-9">
+                        {/* Reactions */}
+                        <AppPostReactions/>
+                    </div>
+                    <div className="col-12 col-lg-3">
+                        {/* Creation Date */}
+                        <AppPostDateCreation cname="text-end"/>
+                    </div>
+                </div>
+
+            </div>
 
         </div>  {/* Fin del Card Body */}
 
@@ -53,7 +65,7 @@ export default function AppCardArticleDetail(props) {
                     <div className="CommentList my-3">
 
                         {/* Contenedor del avatar y el name */}
-                        <AppPostAuthorInfo cname="justify-content-between" estado="full-info"/>
+                        <AppPostAuthorInfo  cname="justify-content-between" estado="full-info"/>
 
                         {/* Contenedor del comment publicado */}
                         <div class={`w-100 p-2 my-2 border rounded`}>
@@ -64,7 +76,7 @@ export default function AppCardArticleDetail(props) {
                     {/* ************* Post Comments ************** */}
                     
                     {/* Contenedor del avatar y el name */}
-                    <AppPostAuthorInfo estado="none"/>
+                    <AppPostAuthorInfo estado=" "/>
 
                         
                      {/* Seccion para Publicar un Comentario    */}
