+++ src/components/Cards/AppCardProductDetail.module.css
