+++ src/components/Cards/AppCardWebinarDetail.module.css
