+++ src/components/Cards/AppPostAuthorInfo.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import AppImage from "./AppImage";
+import AppImage from "../commons/AppImage";
 import AppPostDateCreation from "./AppPostDateCreation";
 import Styles from "./AppPostAuthorInfo.module.css";
